Revert "change `pointId[]` to `long long`"
This reverts commit a552997745160ac47d97998ad7feae176556e823.

+++ src/rgestures.h
@@ -102,7 +102,7 @@ typedef enum {
 typedef struct {
     int touchAction;
     int pointCount;
-    long long pointId[MAX_TOUCH_POINTS];
+    int pointId[MAX_TOUCH_POINTS];
     Vector2 position[MAX_TOUCH_POINTS];
 } GestureEvent;
